pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["[Your Name <you@example.com>]"]
requires-python = ">=3.11"
dependencies = [
    "py-cord>=2.5.0",
    "requests>=2.32.3",
    "Flask>=3.0.0",
    "discord-py>=2.5.2",
]
